fix wrong regex
